backend/requirements.txt
Flask
opencv-python
Werkzeug
flask_cors
python-telegram-bot
requests